submission -included the celo governance proposal part -included the video and image asstes
