+++ governcelo/assets/index.md
@@ -1,2 +1,1 @@
-Images
-Add you logo, other visuals and presentation here.
+#### a video and images demonstrating how the app should look
