+++ governcelo/codebase/governcelodapp/package.json
@@ -9,12 +9,20 @@
   },
   "dependencies": {
     "@octokit/core": "^3.5.1",
+    "@react-native-community/masked-view": "^0.1.11",
+    "@react-navigation/drawer": "^6.1.8",
+    "@react-navigation/native": "^6.0.6",
+    "@react-navigation/stack": "^6.0.11",
     "expo": "~43.0.2",
     "expo-status-bar": "~1.1.0",
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-native": "0.64.3",
-    "react-native-web": "0.17.1"
+    "react-native-web": "0.17.1",
+    "react-native-gesture-handler": "^1.10.3",
+    "react-native-reanimated": "^2.2.3",
+    "react-native-safe-area-context": "^3.3.2",
+    "react-native-screens": "^3.9.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
