+++ governcelo/codebase/index.md
@@ -1,2 +1,13 @@
 codebase directory
 Use this directory to showcase ALL of the code for the hackathon project.
+## Running the app
+Run these from your terminal/cli 
+```git clone https://github.com/ianmunge0/make-crypto-mobile-hackathon.git```
+
+```cd make-crypto-mobile-hackathon/governcelo/codebase/governcelodapp```
+
+```git checkout governcelo```
+
+```yarn```
+
+```expo start```
